Refactor namespaces and cleanup dependencies
Reorganized namespaces and removed unused `using` directives:
- Added `Nalix.Framework.Security.Primitives` to `LZ4Decoder.cs`.
- Removed `Nalix.Framework.Memory.Internal` and other unused directives from `ChaCha20Poly1305.cs`, `Salsa20Poly1305.cs`, `Poly1305.cs`, and `ChaCha20.cs`.

Refactored `MemorySecurity.cs`:
- Changed namespace to `Nalix.Framework.Security.Primitives`.
- Made `MemorySecurity` class `public static` for broader accessibility.
- Removed `InternalsVisibleTo` attributes for test/benchmark assemblies.
- Reformatted `ZeroMemory` method for clarity without altering functionality.

These changes improve modularity, maintainability, and usability.
