+++ src/Nalix.Framework/Security/Symmetric/ChaCha20.cs
@@ -1,7 +1,6 @@
 // Copyright (c) 2025-2026 PPN Corporation. All rights reserved.
 // Licensed under the Apache License, Version 2.0.
 
-using Nalix.Framework.Memory.Internal;
 using Nalix.Framework.Security.Internal;
 using Nalix.Framework.Security.Primitives;
